+++ packages/generator-component/babel.config.js
@@ -16,8 +16,8 @@ module.exports = api => {
     // Hooks into browserslist to provide smart Babel transforms
     // https://babeljs.io/docs/en/babel-preset-env
     presets.push(['@babel/env', {
-        "targets": {
-            "node": "current"
+        targets: {
+            node: 'current'
         }
     }]);