+++ packages/generator-component/CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.2.0
+------------------------------
+*May 12, 2020*
+
+### Changed
+- Updated template to use `vue-test-utils` to v1 and `@vue/cli-plugin-unit-test` to v4.3.1.
+
+### Removed
+- `testMatch` from jest config, as not needed.
+
+
 v0.1.0
 ------------------------------
 *April 21, 2020*